グラフ横スクロール関連のchartjsプラグインを plugins/vue-chart へ移動
